Add in search-by-uid capability for Gitea (>1.7.0)
This should still work on Gitea v1.6.0 - as the uid parameter will
simply be ignored on these versions. The other changes in the
java-gitea-api regarding the use of long are just better
representations than those originally listed in the swagger api.

+++ src/main/java/svnserver/ext/gitea/mapping/GiteaAccess.java
@@ -68,7 +68,7 @@ final class GiteaAccess implements VcsAccess {
               try {
                 final ApiClient apiClient = context.connect();
                 final RepositoryApi repositoryApi = new RepositoryApi(apiClient);
-                final Repository repository = repositoryApi.repoGetByID((int) projectId);
+                final Repository repository = repositoryApi.repoGetByID(projectId);
                 if (!repository.isPrivate()) {
                   // Munge the permissions
                   repository.getPermissions().setAdmin(false);
@@ -88,7 +88,7 @@ final class GiteaAccess implements VcsAccess {
               final ApiClient apiClient = context.connect(userName);
               final RepositoryApi repositoryApi = new RepositoryApi(apiClient);
 
-              final Repository repository = repositoryApi.repoGetByID((int) projectId);
+              final Repository repository = repositoryApi.repoGetByID(projectId);
               return repository;
             } catch (ApiException e) {
               if (e.getCode() == 404) {
